+++ src/arcemu-world/Level3.cpp
@@ -2346,7 +2346,7 @@ bool ChatHandler::HandleCreatureSpawnCommand(const char *args, WorldSession *m_s
 
 	CreatureSpawn * sp = new CreatureSpawn;
 	//sp->displayid = info->DisplayID;
-	gender = (uint8)info->GenerateModelId(&sp->displayid);
+	gender = info->GenerateModelId(&sp->displayid);
  	sp->entry = entry;
 	sp->entry = entry;
 	sp->form = 0;
